+++ Networking/Packets/Outgoing/OtherHitPacket.py
@@ -8,12 +8,12 @@ class OtherHitPacket:
 
     def write(self, writer):
         writer.writeInt32(self.time)
-        writer.writeByte(self.bulletId)
+        writer.writeShort(self.bulletId)
         writer.writeInt32(self.objectId)
         writer.writeInt32(self.targetId)
 
     def read(self, reader):
         self.time = reader.readInt32()
-        self.bulletId = reader.readByte()
+        self.bulletId = reader.readShort()
         self.objectId = reader.readInt32()
         self.targetId = reader.readInt32()
