+++ Networking/Packets/Outgoing/PlayerHitPacket.py
@@ -7,9 +7,9 @@ class PlayerHitPacket:
         self.objectId = 0
 
     def write(self, writer):
-        writer.writeByte(self.bulletId)
+        writer.writeShort(self.bulletId)
         writer.writeInt32(self.objectId)
 
     def read(self, reader):
-        self.bulletId = reader.readByte()
+        self.bulletId = reader.readShort()
         self.objectId = reader.readInt32()
